+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/Internals/HttpPipelineHelper.cs
@@ -385,6 +385,12 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Internals
                     CustomerSdkStatsHelper.TrackDropped(droppedCounter, ResponseStatusCodes.PartialSuccess, "Partial success - non-retriable");
                 }
 
+                // If there is no partialContent, it means all items were either successful or permanently dropped (which does not need to be retried)
+                if (partialContent == null)
+                {
+                    result.ExportResult = ExportResult.Success;
+                }
+
                 return;
             }